Switch to provider class
